fix: add record request/response DTO, add record ErrorDTO - closes #4, closes #5

+++ src/main/java/cat/itacademy/s04/t02/n02/repository/FruitRepository.java
@@ -3,5 +3,5 @@ package cat.itacademy.s04.t02.n02.repository;
 import cat.itacademy.s04.t02.n02.models.Fruit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FruitRepository extends JpaRepository<Fruit, Integer> {
+public interface FruitRepository extends JpaRepository<Fruit, Long> {
 }
